requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Web and API
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0

# Machine Learning and NLP
torch>=1.9.0
transformers>=4.10.0
sentence-transformers>=2.1.0

# Image Processing
opencv-python>=4.5.0
pillow>=8.3.0

# Geospatial
folium==0.14.0
geopy==2.3.0
osmnx>=1.1.0

# Visualization
matplotlib==3.7.2
seaborn>=0.11.0
plotly>=5.3.0

# CLI
click>=8.0.0
rich>=10.0.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Development
black>=21.5b2
flake8>=3.9.0
isort>=5.9.0

# Additional dependencies
Flask-Cors==4.0.0
joblib==1.3.2 